# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "华农掌中行 API",
    "version": "1.0.0",
    "description": "由 back/server.js 与 back/routes/*.js 导出的接口定义（2026-08-14）。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地后端（以 PORT 环境变量为准）"
    }
  ],
...
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "tags": [
          "管理员账号"
        ],
        "summary": "管理员登录",
        "operationId": "post_api_admin_login",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/admin/info": {
      "get": {
        "tags": [
          "管理员账号"
        ],
        "summary": "管理员信息",
        "operationId": "get_api_admin_info",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ]
      }
    },
    "/api/admin/change-password": {
      "post": {
        "tags": [
          "管理员账号"
        ],
        "summary": "修改管理员密码",
        "operationId": "post_api_admin_change_password",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                },
                "required": [
                  "oldPassword",
                  "newPassword"
                ]
              }
            }
          }
        }
      }
    },
    "/api/admin/list": {
      "get": {
        "tags": [
          "管理员账号"
        ],
        "summary": "管理员列表",
        "operationId": "get_api_admin_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ]
      }
    },
    "/api/admin/add": {
      "post": {
        "tags": [
          "管理员账号"
        ],
        "summary": "新增管理员",
        "operationId": "post_api_admin_add",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/admin/update/{id}": {
      "put": {
        "tags": [
          "管理员账号"
        ],
        "summary": "更新管理员",
        "operationId": "put_api_admin_update_id",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "status": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/reset-password/{id}": {
      "post": {
        "tags": [
          "管理员账号"
        ],
        "summary": "重置管理员密码",
        "operationId": "post_api_admin_reset_password_id",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newPassword": {
                    "type": "string"
                  }
                },
                "required": [
                  "newPassword"
                ]
              }
            }
          }
        }
      }
    },
    "/api/admin/delete/{id}": {
      "delete": {
        "tags": [
          "管理员账号"
        ],
        "summary": "删除管理员",
        "operationId": "delete_api_admin_delete_id",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/admin/stats": {
      "get": {
        "tags": [
          "管理员账号"
        ],
        "summary": "后台统计数据",
        "operationId": "get_api_admin_stats",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ]
      }
    },
...
    "/api/admin/flower/routes/list": {
      "get": {
        "tags": [
          "后台业务管理"
        ],
        "summary": "管理端赏花路线列表",
        "operationId": "get_api_admin_flower_routes_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 10
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/admin/flower/routes/add": {
      "post": {
        "tags": [
          "后台业务管理"
        ],
        "summary": "新增赏花路线",
        "operationId": "post_api_admin_flower_routes_add",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "distance": {
                    "type": "number"
                  },
                  "difficulty": {
                    "type": "string"
                  },
                  "spots": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "description": {
                    "type": "string"
                  },
                  "best_time": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "highlights": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "tips": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        }
      }
    },
    "/api/admin/flower/routes/update/{id}": {
      "put": {
        "tags": [
          "后台业务管理"
        ],
        "summary": "更新赏花路线",
        "operationId": "put_api_admin_flower_routes_update_id",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "distance": {
                    "type": "number"
                  },
                  "difficulty": {
                    "type": "string"
                  },
                  "spots": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "description": {
                    "type": "string"
                  },
                  "best_time": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "highlights": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "tips": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/flower/routes/delete/{id}": {
      "delete": {
        "tags": [
          "后台业务管理"
        ],
        "summary": "删除赏花路线",
        "operationId": "delete_api_admin_flower_routes_delete_id",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "用户 JWT，直接传 token"
      },
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "管理员 JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": [
          "code"
        ]
      }
    }
  }
}